User
Check TODO comments for missing implementations

+++ backend/src/main/java/com/motorph/ems/model/User.java
@@ -57,7 +57,7 @@ public class User {
     public static class Role {
         @Id
         @GeneratedValue(strategy = GenerationType.IDENTITY)
-        private Long userRoleId;
+        private int userRoleId;
         private String role;
 
         @Override
